Add ability to query portal for public and private proposals
Adds some configuration to configure a connection to the obs portal and the ability to configure a set of proposal tags that dictate data privacy rules. When ingesting a frame, check if any of these tags exist in a proposal - if they do, use them to set the public date. If not, fall back to original logic contained in PUBLIC_PROPOSALS and PRIVATE_PROPOSALS. Add some tests.